Ignore some module members when looking for its shared functions
Inspecting *all* module attributes lead to premature evaluation of its
properties, which may depend on information not yet known or set up by
the time the discovery process runs. Therefore inspect.getmembers() is
replaced with dir, filtering out well-known special members (e.g.
__init__) and members of Module class as well. The final list of members
to actually inspect is then shorter, and does not contain eval_context.
It still may contain custom properties, as defined by module developer,
but in such a case it's up to the developer to deal with situation when
the module wasn't initialized yet.