Verbose logging goes into separate file (#30)
DEBUG logging reached huge amounts of text, making the examination of common
"debug" log files hard. Also, we're not using "verbose" logging anyway,
so the idea is to use it for logging blobs and long texts - those that
are usually not that interesting in debug log, but might be very useful
when we need to debug weird behavior - API responses, HTTP
communication, command output, eval context gathering.

This patch repurposes our "verbose" logging. Every "verbose" call
creates a tag for the given message, logs this tag with normal DEBUG
level, and then logs the message with VERBOSE level. Special file
handler is added, it ignores any other level except the VERBOSE, and by
combining these two approaches we get a file with VERBOSE messages only,
and a debug output, both connected via the aforementioned tags, making
it possible to take a look into the verbose file for bits that were omitted
from a debug log.