When logging a traceback, show the exception origin on top as well (#44)
When logging a traceback, show the exception origin on top as well

The point of origin is logged btu given the order of stack frames shows
the oldest ones first, the actual place where exception happened is
hidden at the bottom. With large (when logged) local variables, looking
for can be quite unpleasant experience.