When pipeline encouters exception, make a note before proceeding
Previous version of pipeline let the exception bubble up to the tool
which logged it and began "destroy" stage. Current version of pipeline
code catches the exception, begans the "destroy" stage, and returns the
exception to the caller. It works, but the side-effect is that the
exception is logged by the caller (tool in the case of the main
pipeline), which happens at the very end of the pipeline - it is quite
hard to spot in the log *where* the exception happened since its log is
at the end of the output no matter what.

So, this patch adds one `error` call when pipeline encouters the
exception, and slightly changes wording of the final "log the error"
message - we want tool to log it anyway, but that would mean it's logged
twice, hence the final message should have a bit more global-ish sound.