Rename !include to !import-variables (#91)
New name is more fitting, because we're actually not including anything
into the file itself, but importing a set of variables into the
namespace available to macros in the file. The old name also clashed
with the actual !include directive recommended to be used to include
data into the YAML structure, and we will need that directive one day,
therefore getting rid of ambiguity.